+++ hiyoco/src/Components/Compare.jsx
@@ -1,0 +1,8 @@
+export default function Compare()
+{
+    return (
+        <div>
+            I am compare!
+        </div>
+    );
+}
