+++ hiyoco/src/Components/Info.jsx
@@ -1,0 +1,8 @@
+export default function Info()
+{
+    return (
+        <div>
+            I am info!
+        </div>
+    );
+}
